requirements.txt
apache-airflow==2.7.3
pandas==2.1.3
requests==2.31.0
google-api-python-client==2.108.0
boto3==1.33.0
s3fs==2023.10.0